\name{talentlens-package}
\alias{talentlens-package}
\alias{talentlens}
\docType{package}
\title{
Package for TalentLens
}
\description{
Contains functions for getting access to data from a Concerto test (fetchConcertoData), monitoring participants during data collection (monitorParticipants), balancing norm groups (balancer), finding optimal anchor items given a score matrix (findAnchors), scoring responses to items (scoreDataFrame and scoreResponses), estimating theta from a score matrix given item parameters (eap) and generating target proportions for the UK census 2011 (targetUKCensus).

}
\details{
\tabular{ll}{
Package: \tab talentlens\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2016-06-17\cr
License: \tab No licence\cr
}
1. \link{fetchConcertoData}\cr
2. \link{monitorParticipants}\cr
3. \link{balancer}\cr
4. \link{findAnchors}\cr
5. \link{scoreDataFrame}\cr
6. \link{scoreResponses}\cr
7. \link{eap}\cr
8. \link{targetUKCensus}
}
\author{
Morgan Strom
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
