\name{talentlens-package}
\alias{talentlens-package}
\alias{talentlens}
\docType{package}
\title{
Package for TalentLens R&D team
}
\description{
Contains functions for getting access to data from a Concerto test, monitoring participants during data collection, balancing norm groups, finding optimal anchor items given a score matrix, scoring responses to items, estimating theta from a score matrix given item parameters, calculating bagged IRT parameters and generating target proportions according to UK census 2011.

Access information about these functions using the links below, or by typing e.g. '?fetchConcertoData' in the console.

}
\details{
\tabular{ll}{
Package: \tab talentlens\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2016-06-17\cr
License: \tab No licence\cr
}
1. \link{fetchConcertoData}\cr
2. \link{monitorParticipants}\cr
3. \link{balancer}\cr
4. \link{findAnchors}\cr
5. \link{scoreDataFrame}\cr
6. \link{scoreResponses}\cr
7. \link{eap}\cr
8. \link{baggedParams}\cr
9. \link{targetUKCensus}
}
\author{
Morgan Strom
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
